refactor(ds-global): structure the Modal context after the react context standard
