+++ packages/react/ds-global/src/lib/pattern/Modal/index.ts
@@ -1,5 +1,5 @@
 // Generated by @canonical/summon-component:react v0.1.0
 
-export { default as Modal } from "./Modal.js";
+export { default as Modal } from "./Provider.js";
 export type * from "./types.js";
 export { default as withModal } from "./withModal.js";
